A grab bag of misc fixes (#3526)
* Remove last traces of FSTObjectValue

* Fix typos

* Add ToString to SnapshotVersion

* Add support for hashing enums and optionals

+++ Firestore/core/src/firebase/firestore/core/field_filter.cc
@@ -176,7 +176,7 @@ std::string FieldFilter::ToString() const {
 }
 
 size_t FieldFilter::Hash() const {
-  return util::Hash(field_, static_cast<int>(op_), value_rhs_);
+  return util::Hash(field_, op_, value_rhs_);
 }
 
 bool FieldFilter::IsInequality() const {
